backend/requirements.txt
fastapi
uvicorn
tree-sitter==0.21.3
tree-sitter-python==0.21.0
tree-sitter-javascript==0.21.4
tree-sitter-java==0.21.0
chromadb
google-generativeai
python-dotenv
python-multipart
networkx
gitpython
requests